+++ sjasm/devices.cpp
@@ -28,13 +28,17 @@
 
 #include "sjdefs.h"
 
-bool IsZXSpectrumDevice(char *name){
-	if (strcmp(name, "ZXSPECTRUM48") && 
-		strcmp(name, "ZXSPECTRUM128") && 
-		strcmp(name, "ZXSPECTRUM256") && 
-		strcmp(name, "ZXSPECTRUM512") && 
-		strcmp(name, "ZXSPECTRUM1024")) {
-			return false;
+bool IsZXSpectrumDevice(char *name) {
+	if (strcmp(name, "ZXSPECTRUM48") &&
+		strcmp(name, "ZXSPECTRUM128") &&
+		strcmp(name, "ZXSPECTRUM256") &&
+		strcmp(name, "ZXSPECTRUM512") &&
+		strcmp(name, "ZXSPECTRUM1024") &&
+		strcmp(name, "ZXSPECTRUM2048") &&
+		strcmp(name, "ZXSPECTRUM4096") &&
+		strcmp(name, "ZXSPECTRUM8192"))
+	{
+		return false;
 	}
 	return true;
 }
@@ -49,51 +53,89 @@ static void initRegularSlotDevice(CDevice* const device, const int32_t slotSize,
 		device->AddPage(memPtr, slotSize);
 	}
 	for (int i = 0; i < device->SlotsCount; ++i) {
-		device->GetSlot(i)->Page = device->GetPage(initialPages[i]);
+		CDeviceSlot* slot = device->GetSlot(i);
+		slot->Page = device->GetPage(initialPages[i]);
+		slot->InitialPage = slot->Page->Number;
 	}
 	device->SetSlot(device->SlotsCount - 1);
 }
 
-static void initZxLikeDevice(CDevice* const device, int32_t slotSize, int pageCount, const int* const initialPages) {
+static void initZxLikeDevice(
+	CDevice* const device, int32_t slotSize, int pageCount, const int* const initialPages, aint ramtop)
+{
 	initRegularSlotDevice(device, slotSize, 0x10000/slotSize, pageCount, initialPages);
-	// set memory to "USR 0"-like state (for snapshot saving) (works also for ZXN 0x2000 slotSize)
-	int vramPage = (0x2000 == slotSize) ? initialPages[2] : initialPages[1];	// default = second slot page
-	byte* const vramRAM = device->GetPage(vramPage)->RAM;
-	memset(vramRAM + 0x1800, 7*8, 768);
-	memcpy(vramRAM + 0x1C00, BASin48Vars, sizeof(BASin48Vars));
-	byte* const stackRAM = device->GetCurrentSlot()->Page->RAM;	// last slot page is default "stack"
-	memcpy(stackRAM + slotSize - sizeof(BASin48SP), BASin48SP, sizeof(BASin48SP));
+	device->ZxRamTop = (0x5D00 <= ramtop && ramtop <= 0xFFFF) ? ramtop : ZX_RAMTOP_DEFAULT;
+
+	// set memory to state like if (for snapshot saving):
+		// CLEAR ZxRamTop (0x5D5B default) : LOAD "bin" CODE : ... USR start_adr
+	aint adr;
+	// ULA attributes: INK 0 : PAPER 7 : FLASH 0 : BRIGTH 0
+	for (adr = 0x5800; adr < 0x5B00; ++adr) device->Poke(adr, 7*8);
+
+	// set UDG data at ZX_UDG_ADR (0xFF58)
+	adr = ZX_UDG_ADR;
+	for (byte value : ZX_UDG_DATA) device->Poke(adr++, value);
+
+	// ZX SYSVARS at 0x5C00
+	adr = ZX_SYSVARS_ADR;
+	for (byte value : ZX_SYSVARS_DATA) device->Poke(adr++, value);
+	// set RAMTOP sysvar
+	device->Poke(ZX_SYSVARS_ADR+0xB2, device->ZxRamTop & 0xFF);
+	device->Poke(ZX_SYSVARS_ADR+0xB3, (device->ZxRamTop>>8) & 0xFF);
+
+	// set STACK data (without the "start" address)
+	adr = device->ZxRamTop + 1 - sizeof(ZX_STACK_DATA);
+	// set ERRSP sysvar to beginning of the fake stack
+	device->Poke(ZX_SYSVARS_ADR+0x3D, adr & 0xFF);
+	device->Poke(ZX_SYSVARS_ADR+0x3E, (adr>>8) & 0xFF);
+	for (byte value : ZX_STACK_DATA) device->Poke(adr++, value);
 }
 
-static void DeviceZXSpectrum48(CDevice **dev, CDevice *parent) {		// add new device
+static void DeviceZXSpectrum48(CDevice **dev, CDevice *parent, aint ramtop) {
 	*dev = new CDevice("ZXSPECTRUM48", parent);
 	const int initialPages[] = {0, 1, 2, 3};
-	initZxLikeDevice(*dev, 0x4000, 4, initialPages);
+	initZxLikeDevice(*dev, 0x4000, 4, initialPages, ramtop);
 }
 
 const static int initialPagesZx128[] = {7, 5, 2, 0};
 
-static void DeviceZXSpectrum128(CDevice **dev, CDevice *parent) {		// add new device
+static void DeviceZXSpectrum128(CDevice **dev, CDevice *parent, aint ramtop) {
 	*dev = new CDevice("ZXSPECTRUM128", parent);
-	initZxLikeDevice(*dev, 0x4000, 8, initialPagesZx128);
+	initZxLikeDevice(*dev, 0x4000, 8, initialPagesZx128, ramtop);
 }
 
-static void DeviceZXSpectrum256(CDevice **dev, CDevice *parent) {		// add new device
+static void DeviceZXSpectrum256(CDevice **dev, CDevice *parent, aint ramtop) {
 	*dev = new CDevice("ZXSPECTRUM256", parent);
-	initZxLikeDevice(*dev, 0x4000, 16, initialPagesZx128);
+	initZxLikeDevice(*dev, 0x4000, 16, initialPagesZx128, ramtop);
 }
 
-static void DeviceZXSpectrum512(CDevice **dev, CDevice *parent) {		// add new device
+static void DeviceZXSpectrum512(CDevice **dev, CDevice *parent, aint ramtop) {
 	*dev = new CDevice("ZXSPECTRUM512", parent);
-	initZxLikeDevice(*dev, 0x4000, 32, initialPagesZx128);
+	initZxLikeDevice(*dev, 0x4000, 32, initialPagesZx128, ramtop);
 }
 
-static void DeviceZXSpectrum1024(CDevice **dev, CDevice *parent) {		// add new device
+static void DeviceZXSpectrum1024(CDevice **dev, CDevice *parent, aint ramtop) {
 	*dev = new CDevice("ZXSPECTRUM1024", parent);
-	initZxLikeDevice(*dev, 0x4000, 64, initialPagesZx128);
+	initZxLikeDevice(*dev, 0x4000, 64, initialPagesZx128, ramtop);
 }
 
-static void DeviceZxSpectrumNext(CDevice **dev, CDevice *parent) {
+static void DeviceZXSpectrum2048(CDevice **dev, CDevice *parent, aint ramtop) {
+	*dev = new CDevice("ZXSPECTRUM2048", parent);
+	initZxLikeDevice(*dev, 0x4000, 128, initialPagesZx128, ramtop);
+}
+
+static void DeviceZXSpectrum4096(CDevice **dev, CDevice *parent, aint ramtop) {
+	*dev = new CDevice("ZXSPECTRUM4096", parent);
+	initZxLikeDevice(*dev, 0x4000, 256, initialPagesZx128, ramtop);
+}
+
+static void DeviceZXSpectrum8192(CDevice **dev, CDevice *parent, aint ramtop) {
+	*dev = new CDevice("ZXSPECTRUM8192", parent);
+	initZxLikeDevice(*dev, 0x4000, 512, initialPagesZx128, ramtop);
+}
+
+static void DeviceZxSpectrumNext(CDevice **dev, CDevice *parent, aint ramtop) {
+	if (ramtop) Warning("ZXN device doesn't init memory in any way (RAMTOP is ignored)");
 	if (Options::IsI8080) Error("Can't use ZXN device while in i8080 assembling mode.", line, FATAL);
 	if (Options::IsLR35902) Error("Can't use ZXN device while in Sharp LR35902 assembling mode.", line, FATAL);
 	*dev = new CDevice("ZXSPECTRUMNEXT", parent);
@@ -107,37 +149,42 @@ static void DeviceZxSpectrumNext(CDevice **dev, CDevice *parent) {
 	}
 }
 
-int SetDevice(char *id) {
+int SetDevice(char *id, const aint ramtop) {
 	CDevice** dev;
-	CDevice* parent;
+	CDevice* parent = nullptr;
 
 	if (!id || cmphstr(id, "none")) {
 		DeviceID = 0; return true;
 	}
 
-	if (!DeviceID || strcmp(DeviceID, id)) {
+	if (!DeviceID || strcmp(DeviceID, id)) {	// different device than current, change to it
 		DeviceID = 0;
 		dev = &Devices;
-		parent = 0;
 		// search for device
 		while (*dev) {
 			parent = *dev;
 			if (!strcmp(parent->ID, id)) break;
 			dev = &(parent->Next);
 		}
-		if (NULL == *dev) {		// device not found
+		if (nullptr == (*dev)) {	// device not found
 			if (cmphstr(id, "zxspectrum48")) {			// must be lowercase to catch both cases
-				DeviceZXSpectrum48(dev, parent);
+				DeviceZXSpectrum48(dev, parent, ramtop);
 			} else if (cmphstr(id, "zxspectrum128")) {
-				DeviceZXSpectrum128(dev, parent);
+				DeviceZXSpectrum128(dev, parent, ramtop);
 			} else if (cmphstr(id, "zxspectrum256")) {
-				DeviceZXSpectrum256(dev, parent);
+				DeviceZXSpectrum256(dev, parent, ramtop);
 			} else if (cmphstr(id, "zxspectrum512")) {
-				DeviceZXSpectrum512(dev, parent);
+				DeviceZXSpectrum512(dev, parent, ramtop);
 			} else if (cmphstr(id, "zxspectrum1024")) {
-				DeviceZXSpectrum1024(dev, parent);
+				DeviceZXSpectrum1024(dev, parent, ramtop);
+			} else if (cmphstr(id, "zxspectrum2048")) {
+				DeviceZXSpectrum2048(dev, parent, ramtop);
+			} else if (cmphstr(id, "zxspectrum4096")) {
+				DeviceZXSpectrum4096(dev, parent, ramtop);
+			} else if (cmphstr(id, "zxspectrum8192")) {
+				DeviceZXSpectrum8192(dev, parent, ramtop);
 			} else if (cmphstr(id, "zxspectrumnext")) {
-				DeviceZxSpectrumNext(dev, parent);
+				DeviceZxSpectrumNext(dev, parent, ramtop);
 			} else {
 				return false;
 			}
@@ -146,6 +193,9 @@ int SetDevice(char *id) {
 		Device = (*dev);
 		DeviceID = Device->ID;
 		Device->CheckPage(CDevice::CHECK_RESET);
+	}
+	if (ramtop && Device->ZxRamTop && ramtop != Device->ZxRamTop) {
+		Warning("[DEVICE] this device was already opened with different RAMTOP value");
 	}
 	return true;
 }
@@ -159,7 +209,7 @@ char* GetDeviceName() {
 }
 
 CDevice::CDevice(const char *name, CDevice *parent)
-	: Next(NULL), SlotsCount(0), PagesCount(0), Memory(nullptr), CurrentSlot(0) {
+	: Next(NULL), SlotsCount(0), PagesCount(0), Memory(nullptr), ZxRamTop(0), CurrentSlot(0) {
 	ID = STRDUP(name);
 	if (parent) parent->Next = this;
 	for (auto & slot : Slots) slot = NULL;
@@ -175,10 +225,12 @@ CDevice::~CDevice() {
 }
 
 void CDevice::AddSlot(int32_t adr, int32_t size) {
+	if (MAX_SLOT_N == SlotsCount) ErrorInt("Can't add more slots, already at max", MAX_SLOT_N, FATAL);
 	Slots[SlotsCount++] = new CDeviceSlot(adr, size);
 }
 
 void CDevice::AddPage(byte* memory, int32_t size) {
+	if (MAX_PAGE_N == PagesCount) ErrorInt("Can't add more pages, already at max", MAX_PAGE_N, FATAL);
 	Pages[PagesCount] = new CDevicePage(memory, size, PagesCount);
 	PagesCount++;
 }
@@ -216,8 +268,8 @@ int CDevice::GetPageOfA16(int32_t address) {
 
 void CDevice::CheckPage(const ECheckPageLevel level) {
 	// fake DISP address gets auto-wrapped FFFF->0 (with warning only)
-	// ("no emit" to catch before labels are defined, although "emit" sounds more logical)
-	if (PseudoORG && CHECK_NO_EMIT == level && 0x10000 <= CurAddress) {
+	// only with "emit" mode, labels may get the value 0x10000 before the address gets truncated
+	if (PseudoORG && CHECK_NO_EMIT != level && 0x10000 <= CurAddress) {
 		if (LASTPASS == pass) {
 			char buf[64];
 			SPRINTF1(buf, 64, "RAM limit exceeded 0x%X by DISP", (unsigned int)CurAddress);
@@ -313,14 +365,59 @@ int32_t CDevice::GetMemoryOffset(int page, int32_t offset) const {
 	return offset + page * Pages[0]->Size;
 }
 
+void CDevice::Poke(aint z80adr, byte value) {
+	// write byte into device memory with current page-mapping
+	const int adrPage = GetPageOfA16(z80adr);
+	if (-1 == adrPage) return;		// silently ignore invalid address
+	CDevicePage* page = GetPage(adrPage);
+	page->RAM[z80adr & (page->Size-1)] = value;
+}
+
 CDevicePage::CDevicePage(byte* memory, int32_t size, int number)
 	: Size(size), Number(number), RAM(memory) {
 	if (nullptr == RAM) Error("No memory defined", nullptr, FATAL);
 }
 
 CDeviceSlot::CDeviceSlot(int32_t adr, int32_t size)
-	: Address(adr), Size(size), Page(NULL), Option(O_NONE) {
+	: Address(adr), Size(size), Page(NULL), InitialPage(-1), Option(O_NONE) {
 }
 
 CDeviceSlot::~CDeviceSlot() {
 }
+
+const unsigned char ZX_SYSVARS_DATA[] = {
+	0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0x00, 0x14, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x01, 0x00, 0x06, 0x00, 0x0b, 0x00, 0x01, 0x00, 0x01, 0x00, 0x06, 0x00, 0x10, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x3c, 0x40, 0x00, 0xff, 0xcc, 0x01, 0x54, 0xff, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x38, 0x00, 0x00, 0xcb, 0x5c, 0x00, 0x00, 0xb6,
+	0x5c, 0xb6, 0x5c, 0xcb, 0x5c, 0x00, 0x00, 0xca, 0x5c, 0xcc, 0x5c, 0xcc, 0x5c, 0xcc, 0x5c, 0x00,
+	0x00, 0xce, 0x5c, 0xce, 0x5c, 0xce, 0x5c, 0x00, 0x92, 0x5c, 0x10, 0x02, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x58, 0xff, 0x00, 0x00, 0x21,
+	0x5b, 0x00, 0x21, 0x17, 0x00, 0x40, 0xe0, 0x50, 0x21, 0x18, 0x21, 0x17, 0x01, 0x38, 0x00, 0x38,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x57, 0xff, 0xff, 0xff, 0xf4, 0x09, 0xa8, 0x10, 0x4b, 0xf4, 0x09, 0xc4, 0x15, 0x53,
+	0x81, 0x0f, 0xc4, 0x15, 0x52, 0xf4, 0x09, 0xc4, 0x15, 0x50, 0x80, 0x80, 0x0d, 0x80, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+
+const unsigned char ZX_STACK_DATA[] = {
+	0x03, 0x13, 0x00, 0x3e
+};
+
+const unsigned char ZX_UDG_DATA[168] = {
+	0x00, 0x3c, 0x42, 0x42, 0x7e, 0x42, 0x42, 0x00, 0x00, 0x7c, 0x42, 0x7c, 0x42, 0x42, 0x7c, 0x00,
+	0x00, 0x3c, 0x42, 0x40, 0x40, 0x42, 0x3c, 0x00, 0x00, 0x78, 0x44, 0x42, 0x42, 0x44, 0x78, 0x00,
+	0x00, 0x7e, 0x40, 0x7c, 0x40, 0x40, 0x7e, 0x00, 0x00, 0x7e, 0x40, 0x7c, 0x40, 0x40, 0x40, 0x00,
+	0x00, 0x3c, 0x42, 0x40, 0x4e, 0x42, 0x3c, 0x00, 0x00, 0x42, 0x42, 0x7e, 0x42, 0x42, 0x42, 0x00,
+	0x00, 0x3e, 0x08, 0x08, 0x08, 0x08, 0x3e, 0x00, 0x00, 0x02, 0x02, 0x02, 0x42, 0x42, 0x3c, 0x00,
+	0x00, 0x44, 0x48, 0x70, 0x48, 0x44, 0x42, 0x00, 0x00, 0x40, 0x40, 0x40, 0x40, 0x40, 0x7e, 0x00,
+	0x00, 0x42, 0x66, 0x5a, 0x42, 0x42, 0x42, 0x00, 0x00, 0x42, 0x62, 0x52, 0x4a, 0x46, 0x42, 0x00,
+	0x00, 0x3c, 0x42, 0x42, 0x42, 0x42, 0x3c, 0x00, 0x00, 0x7c, 0x42, 0x42, 0x7c, 0x40, 0x40, 0x00,
+	0x00, 0x3c, 0x42, 0x42, 0x52, 0x4a, 0x3c, 0x00, 0x00, 0x7c, 0x42, 0x42, 0x7c, 0x44, 0x42, 0x00,
+	0x00, 0x3c, 0x40, 0x3c, 0x02, 0x42, 0x3c, 0x00, 0x00, 0xfe, 0x10, 0x10, 0x10, 0x10, 0x10, 0x00,
+	0x00, 0x42, 0x42, 0x42, 0x42, 0x42, 0x3c, 0x00
+};