+++ sjasm/sjasm.cpp
@@ -50,6 +50,7 @@ void PrintHelp() {
 	_COUT "  --zxnext[=cspect]        Enable ZX Spectrum Next Z80 extensions (Z80N)" _ENDL;
 	_COUT "  --i8080                  Limit valid instructions to i8080 only (+ no fakes)" _ENDL;
 	_COUT "  --lr35902                Sharp LR35902 CPU instructions mode (+ no fakes)" _ENDL;
+	_COUT "  --outprefix=<path>       Prefix for save/output/.. filenames in directives" _ENDL;
 	_COUT "  -i<path> or -I<path> or --inc=<path> ( --inc without \"=\" to empty the list)" _ENDL;
 	_COUT "                           Include path (later defined have higher priority)" _ENDL;
 	_COUT "  --lst[=<filename>]       Save listing to <filename> (<source>.lst is default)" _ENDL;
@@ -64,10 +65,10 @@ void PrintHelp() {
 	_COUT "  --nologo                 Do not show startup message" _ENDL;
 	_COUT "  --msg=[all|war|err|none|lst|lstlab]" _ENDL;
 	_COUT "                           Stderr messages verbosity (\"all\" is default)" _ENDL;
-	_COUT "  --fullpath               Show full path to error file" _ENDL;
+	_COUT "  --fullpath               Show full path to file in errors" _ENDL;
 	_COUT " Other:" _ENDL;
 	_COUT "  -D<NAME>[=<value>]       Define <NAME> as <value>" _ENDL;
-	_COUT "  -                        Reads STDIN as source (no other sourcefile allowed)" _ENDL;
+	_COUT "  -                        Reads STDIN as source (even in between regular files)" _ENDL;
 	_COUT "  --longptr                No device: program counter $ can go beyond 0x10000" _ENDL;
 	_COUT "  --reversepop             Enable reverse POP order (as in base SjASM version)" _ENDL;
 	_COUT "  --dirbol                 Enable directives from the beginning of line" _ENDL;
@@ -77,6 +78,7 @@ void PrintHelp() {
 }
 
 namespace Options {
+	char OutPrefix[LINEMAX] = {0};
 	char SymbolListFName[LINEMAX] = {0};
 	char ListingFName[LINEMAX] = {0};
 	char ExportFName[LINEMAX] = {0};
@@ -101,7 +103,6 @@ namespace Options {
 	bool ShowVersion = false;
 	bool NoDestinationFile = true;		// no *.out file by default
 	SSyntax syx, systemSyntax;
-	bool SourceStdIn = false;
 	bool IsI8080 = false;
 	bool IsLR35902 = false;
 	bool IsLongPtr = false;
@@ -169,10 +170,30 @@ const char* fileNameFull = nullptr, * fileName = nullptr;	//fileName is either f
 char* lp, line[LINEMAX], temp[LINEMAX], ErrorLine[LINEMAX2], ErrorLine2[LINEMAX2], * bp;
 char sline[LINEMAX2], sline2[LINEMAX2], * substitutedLine, * eolComment, ModuleName[LINEMAX];
 
-char SourceFNames[128][MAX_PATH];
-static int SourceFNamesCount = 0;
-std::vector<std::string> openedFileNames(256);
-std::vector<char> stdin_log;
+SSource::SSource(SSource && src) {	// move constructor, "pick" the stdin pointer
+	memcpy(fname, src.fname, MAX_PATH);
+	stdin_log = src.stdin_log;
+	src.fname[0] = 0;
+	src.stdin_log = nullptr;
+}
+
+SSource::SSource(const char* newfname) : stdin_log(nullptr) {
+	STRNCPY(fname, MAX_PATH, newfname, MAX_PATH-1);
+	fname[MAX_PATH-1] = 0;
+}
+
+SSource::SSource(int) {
+	fname[0] = 0;
+	stdin_log = new stdin_log_t();
+	stdin_log->reserve(50*1024);
+}
+
+SSource::~SSource() {
+	if (stdin_log) delete stdin_log;
+}
+
+std::vector<SSource> sourceFiles;
+std::vector<std::string> openedFileNames;
 
 int ConvertEncoding = ENCWIN;
 
@@ -436,6 +457,7 @@ namespace Options {
 				} else if (!strcmp(opt, "sld") && !val[0]) {
 					IsDefaultSldName = true;
 				} else if (
+					CheckAssignmentOption("outprefix", OutPrefix, LINEMAX) ||
 					CheckAssignmentOption("sym", SymbolListFName, LINEMAX) ||
 					CheckAssignmentOption("lst", ListingFName, LINEMAX) ||
 					CheckAssignmentOption("exp", ExportFName, LINEMAX) ||
@@ -471,8 +493,8 @@ namespace Options {
 						_CERR "No parameters found in " _CMDL arg _ENDL;
 					}
 				} else if (!doubleDash && 0 == opt[0]) {
-					SourceStdIn = true;		// only single "-" was on command line = source STDIN
-					stdin_log.reserve(100000);	// reserve 100k bytes for a start
+					// only single "-" was on command line = source STDIN
+					sourceFiles.push_back(SSource(1));		// special constructor for stdin input
 				} else {
 					_CERR "Unrecognized option: " _CMDL arg _ENDL;
 				}
@@ -534,8 +556,10 @@ int main(int argc, char **argv) {
 #endif
 	char buf[MAX_PATH];
 	int base_encoding;
-	char* p;
 	const char* logo = "SjASMPlus Z80 Cross-Assembler v" VERSION " (https://github.com/z00m128/sjasmplus)";
+
+	sourceFiles.reserve(32);
+	openedFileNames.reserve(64);
 
 	CHECK_UNIT_TESTS		// UnitTest++ extra handling in specially built executable
 
@@ -554,7 +578,7 @@ int main(int argc, char **argv) {
 	long dwStart = GetTickCount();
 
 	// get current directory
-	GetCurrentDirectory(MAX_PATH, buf);
+	SJ_GetCurrentDirectory(MAX_PATH, buf);
 	CurrentDirectory = buf;
 
 	Options::COptionsParser optParser;
@@ -575,8 +599,8 @@ int main(int argc, char **argv) {
 		int i = 0;
 		while (parsedOptsArray[i]) {
 			optParser.GetOptions(parsedOptsArray, i);
-			if (!parsedOptsArray[i] || 128 <= SourceFNamesCount) break;
-			STRCPY(SourceFNames[SourceFNamesCount++], MAX_PATH-32, parsedOptsArray[i++]);
+			if (!parsedOptsArray[i]) break;
+			sourceFiles.push_back(SSource(parsedOptsArray[i++]));
 		}
 	}
 
@@ -584,8 +608,8 @@ int main(int argc, char **argv) {
 	if (argc > 1) {
 		while (argv[i]) {
 			optParser.GetOptions(argv, i);
-			if (!argv[i] || 128 <= SourceFNamesCount) break;
-			STRCPY(SourceFNames[SourceFNamesCount++], MAX_PATH-32, argv[i++]);
+			if (!argv[i]) break;
+			sourceFiles.push_back(SSource(argv[i++]));
 		}
 		if (Options::IsDefaultListingName && Options::ListingFName[0]) {
 			Error("Using both  --lst  and  --lst=<filename>  is not possible.", NULL, FATAL);
@@ -615,7 +639,7 @@ int main(int argc, char **argv) {
 		}
 		// otherwise the full logo was already printed
 		// now check if there were some sources to assemble, if NOT, exit with "OK"!
-		if (!SourceFNames[0][0] && !Options::SourceStdIn) exit(0);
+		if (0 == sourceFiles.size()) exit(0);
 	}
 
 #ifdef USE_LUA
@@ -631,37 +655,15 @@ int main(int argc, char **argv) {
 #endif //USE_LUA
 
 	// exit with error if no input file were specified
-	if (!SourceFNames[0][0] && !Options::SourceStdIn) {
+	if (0 == sourceFiles.size()) {
 		if (Options::OutputVerbosity <= OV_ERROR) {
 			_CERR "No inputfile(s)" _ENDL;
 		}
 		exit(1);
 	}
-	// verify for STDIN input there is no other file specified + create empty name signaling STDIN
-	if (Options::SourceStdIn) {
-		if (0 < SourceFNamesCount) {	// list of explicit input files must be empty with `-` option
-			if (Options::OutputVerbosity <= OV_ERROR) {
-				_CERR "Don't add input file when STDIN option is specified." _ENDL;
-			}
-			exit(1);
-		}
-		// stdin itself has empty filename
-		SourceFNames[SourceFNamesCount++][0] = 0;
-		// but fake output name if not selected explicitly
-		if (!Options::DestinationFName[0]) STRCPY(Options::DestinationFName, LINEMAX, "asm.out");
-	}
 
 	// create default output name, if not specified
-	if (!Options::DestinationFName[0]) {
-		STRCPY(Options::DestinationFName, LINEMAX, SourceFNames[0]);
-		if (!(p = strchr(Options::DestinationFName, '.'))) {
-			p = Options::DestinationFName;
-		} else {
-			*p = 0;
-		}
-		STRCAT(p, LINEMAX-(p-Options::DestinationFName), ".out");
-	}
-
+	ConstructDefaultFilename(Options::DestinationFName, LINEMAX, ".out");
 	base_encoding = ConvertEncoding;
 
 	// init some vars
@@ -680,10 +682,10 @@ int main(int argc, char **argv) {
 			OpenSld();
 		}
 
-		for (i = 0; i < SourceFNamesCount; i++) {
+		for (SSource & src : sourceFiles) {
 			IsRunning = 1;
 			ConvertEncoding = base_encoding;
-			OpenFile(SourceFNames[i]);
+			OpenFile(src.fname, false, src.stdin_log);
 		}
 
 		if (PseudoORG) {