+++ sjasm/tables.h
@@ -51,6 +51,7 @@ int GetLocalLabelValue(char*& op, aint& val);
 
 constexpr int LABEL_PAGE_UNDEFINED = -1;
 constexpr int LABEL_PAGE_ROM = 0x7F00;			// must be minimum of special values (but positive)
+constexpr int LABEL_PAGE_OUT_OF_BOUNDS = 0x7F01;	// label is defined, but not within Z80 address space
 
 class CLabelTableEntry {
 public: